reorg(0254): extract venue-naming helpers to flat _venue_naming.py
summarize_core_venues.py was dual-role: a Phase-2 entry point that renders
the core-venue tables, but also imported by compute_venue_concentration.py
and export_core_venues_markdown.py for its canonical_venue / venue_type
string-classification helpers. That script→script import blocks the wave-1
move (a moved entry point breaks the importer's `from summarize_core_venues
import ...`).

Relocate _VENUE_RULES + canonical_venue / venue_type / institution_group
verbatim into a neutral flat _venue_naming.py (the 0250 pattern) and repoint
both importers. Output is byte-identical by construction — the rules and
function bodies are moved unchanged (verified on 15 representative venue
strings across all three helpers). summarize_core_venues.py is now a pure
entry point, free to move in wave-1 (0257).

Ticket-ref: tickets/0254-reorg-code-entry-point-vs-library-classi.erg
